+++ src/main.cpp
@@ -17,10 +17,11 @@ SDL_Renderer* gRenderer;
 
 Mix_Music* sMusic = NULL;   
 
-int testRand = randNum(20,60);
+int Operator;
 
 void initialise();
 void loadAssets();
+int getRandOperator();
 void quit();
 
 // Texture class
@@ -98,6 +99,7 @@ int WinMain(int argc, char* argv[]) {
     bool stop = false;
     SDL_Event e;
 
+    
 
     while (!stop) {
         while (SDL_PollEvent(&e) != 0) {
@@ -105,6 +107,24 @@ int WinMain(int argc, char* argv[]) {
                 stop = true;
             }
         }
+
+        // Game logic
+        Operator = randNum(1,4);
+        switch (Operator) {
+            case 1:
+                printf("+");
+                break;
+            case 2:
+                printf("-");
+                break;
+            case 3:
+                printf("*");
+                break;
+            case 4:
+                printf("/");
+                break;
+            }
+
         // Graphical rendering
         SDL_SetRenderDrawColor(gRenderer, 255, 255, 255, 255);
         SDL_RenderClear(gRenderer);
@@ -127,13 +147,14 @@ void initialise() {
     Mix_OpenAudio(44100, MIX_DEFAULT_FORMAT, 2, 2048);
     gWindow = SDL_CreateWindow(("MathOrDeath " + VERSION).c_str(), SDL_WINDOWPOS_CENTERED, SDL_WINDOWPOS_CENTERED, SCR_WIDTH, SCR_HEIGHT, SDL_WINDOW_SHOWN);
     gRenderer = SDL_CreateRenderer(gWindow, -1, SDL_RENDERER_ACCELERATED | SDL_RENDERER_PRESENTVSYNC); // Accelerated with VSync activated
+    SDL_SetWindowFullscreen(gWindow,SDL_WINDOW_FULLSCREEN_DESKTOP);
 }
 
 void loadAssets() {
-    gTestBackground.loadFromFile("res/img/misc/background-0001.png");   
+    gTestBackground.loadFromFile("res/img/misc/background-0002.png");   
     sMusic = Mix_LoadMUS("res/sfx/music/test.ogg");
 }
-    	
+
 void quit() {
 
     // Graphical elements
@@ -151,4 +172,6 @@ void quit() {
     */  
     IMG_Quit();
     SDL_Quit();
+
+    Operator = NULL;
 }